Fix BDD tests for dynatable elements, broken in earlier commit
Commit 601cc5822c782a2652a2725db4ed301770836430 changed
PageObject::App::Roles::Dynatable to look only at visible dynatable
elements on the page, using an `is_displayed` test, but this
failed for testing with Firefox.

This patch is a simplification using just xpath to find the first
visible dynatable element, which works for both Chrome and Firefox
browser tests.