Fix Remark field doesn't populate when editing Bank Account
Clicking on an existing bank account should populate the corresponding
input fields with the existing values, allowing them to be edited and
used as a base for creatng a new bank account.

The `Remark` field was not being populated - fixed using the same
design pattern as applied to editing `Addresses`.

This change removes the need to round-trip field values to the
server as url parameters. Instead just the bank account `id` is passed.