Add BDD tests for `editing` a bank account
This adds tests for clicking on a bank account and populating the
input fields with that account's attributes. This can then be used
as the basis for creating a new bank account.

There isn't actually a way to update an existing bank account in
the UI. That's for the future...