Cleanup on removal of <lsmb-date> Web Component from DOM
We were creating a `<span>` wrapper element in the `connectedCallback()`
method when adding the `<lsmb-date>` component, but not reversing this
when removing the component.

This patch adds a `disconnectedCallback()` method, which cleans-up.

Without this, the position of label and date selector elements were
messed up when moving the element within the DOM, as the `<label>`
element was being destroyed and recreated, but the widget container
was not.