Clean up when removing <lsmb-button> element from DOM
When removed/disconnected from the DOM, the `<lsmb-button>` element
`disconnectedCallback()` method should fully reverse actions taken by the
`connectedCallback()` method when it was added/connected.

This patch adds the missing reversal steps.

Without this, parts of the button element are duplicated if the element
is relocated within the DOM, leading to a button within a button.