Fix contact__search by country
This patch fixes the `contact__search` country parameter to filer
results by the entity's short country name.

Previously, entity search by country was broken. It only returned entities
with a credit account, while its filtering tested only that the
specified country text was part of a valid country name, not
that it was actually associated with the returned results.

We've now introduced a drop-down select to pick the country filter,
which returns short country name as a value, so we need search only by this,
not the full country name.

Entities, either people or companies, are required to be associated
with a country. It is against this association that we need to filter -
answering a question such as: "Show me all the Canadian companies".