1.12 Fix contact__search by country
This is a mininam fix for `contact__search` country parameter to filer
results by the entity's short country name.

It allows entering of either a short country code, or the initial
letters of a full country name into the Contact Search filter form, to
return people or companies from a matching country.

Previously, entity search by country was broken. It only returned entities
with a credit account, while its filtering tested only that the
specified country text was part of a valid country name, not
that it was actually associated with the returned results.