Fix database change `notify-import-financial-report-templates.sql`
When upgrading databases prior to 1.8, there's a check relating to
balance sheet and income statement templates having been moved to the
database.

If these templates have been loaded into the database, the check
passes without issue. If not, a notification should be generated.

The notification was defined with malformed json, causing the
following fatal error if triggered during setup:

`, or } expected while parsing object/hash, at character offset 176
(before "System > Templates" ...") at
/srv/ledgersmb/lib/LedgerSMB/Database/Change.pm line 336.`

The bug is corrected in this PR.