Fix 1.13 missing-fkeys db change
Part of this change ensures that `language code` references in various
tables have a corresponding entry in the `languages` table, before
enforcing foreign keys.

The `oe`, `ar` and `ap` table schemas permit a NULL `language_code`. Such
rows are now ignored by this change script.

Previously, the change script tried to create a corresponding NULL record
in the `language` table, which is forbidden by a NOT NULL constraint and
would fail, causing a fatal error.