Fix 1.13 db change to drop notes table
The database change `sql/changes/1.13/drop-inheritance-notes-translations.sql`
was failing to drop the `notes` table as it was still referenced by the
`eca__list_notes(integer)` function.

This changes adds `CASCADE` to the drop command.

This also temporarily drops the `eca__list_notes(integer)` function,
which gets replaced later in the upgrade process when modules are reloaded.