Fix orphan data separators in report filenames
For report file exports (csv, ods, pdf, etc), the auto-generated
filename can include the date range for the report. The `-` and `_`
date separators were being included regardless of whether a
"from date" or "to date" applied.

This PR fixed `_-` being appended to the filenames of reports
with no date range filter specified.