Constrain languages.code database field
Enforce the same validation check at database level as is applied by
the web api and ui. Without this check it is possible for the database
to hold values for `language.code` which will cause the ui to error.

Clean-up legacy empty-string language code values in ar, ap and oe tables,
replacing them with NULL values as a precursor to enforcing referential
integrity with the language table.