Ensure Dojo parsing after loading the same route a second time
When loading the same (menu) route twice (e.g. Preferences), the
controls stop working, because they are unloaded from Dojo. They
are supposed to be reloaded, but Dojo skips that step. Apparently,
it knows the DOM subtree has already been parsed. To ensure correct
handling, clear the content and reset the same content after Vue
has established the empty content as new content, causing both
Vue as well as Dojo to see the route's content as new and in need
of processing.