Simplifying Use Overpayment Code
The code is extremely complex for what it does.

PLEASE NOTE:  this does mess with localization with two strings in
warnings in the use overpayment workflow.  However not only was this
workflow broken but the use of the translation API was also broken and
therefore I do not believe this should prevent backporting.