Split the Jest config into per-project and global keys
Explicitly merge keys that are allowed as per-project into the project configurations,
with shared definition for keys that we want to be common. Rumor has it that Jest does
the merging operation in a way that babel-jest and vue-jest *may* fail to see the correct
(ie. merged) config, which was part of the problem getting Quasar to work.